+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/AllTokens.vue
@@ -53,7 +53,6 @@
       @change="setTheme"
     />
   </div>
-  <token-tree :node="filteredTokens" :category="null" :level="2" />
   <div
     v-if="noSearchResults"
     class="d-d-flex d-fl-center d-p16 d-gg4 d-fc-tertiary d-fs-300"
@@ -63,12 +62,13 @@
       &OpenCurlyDoubleQuote;{{ searchCriteria }}&CloseCurlyDoubleQuote;
     </strong>
   </div>
+  <token-tree v-else :node="filteredTokens" :category="null" :level="2" />
 </template>
 
 <script setup>
 import tokensJson from '@dialpad/dialtone-tokens/dist/doc.json';
-import { getComposedTypographyTokens, getComposedShadowTokens } from '../common/token-utilities';
-import { CATEGORY_MAP, SUBCATEGORY_MAP, FORMAT_MAP, THEMES, getTokensStructure } from '../common/constants';
+import { getComposedTypographyTokens, getComposedShadowTokens, addTokenToStructure } from '../common/token-utilities';
+import { FORMAT_MAP, THEMES, getTokensStructure } from '../common/constants';
 import TokenTree from '@baseComponents/TokenTree.vue';
 import { capitalize, computed, ref, onBeforeMount } from 'vue';
 import { debounce } from '../common/utilities';
@@ -97,22 +97,21 @@ const formatSelectMenuOptions = computed(() => {
 });
 
 const updateHeaders = () => {
-  filteredHeaders.value = Object.keys(filteredTokens.value).map(category => {
-    const subcategories = Array.isArray(filteredTokens.value[category])
-      ? []
-      : Object.keys(filteredTokens.value[category]).reduce((accum, curr) => {
-        if (curr !== '_children') {
-          accum.push({
-            title: capitalize(curr),
-            level: 2,
-            slug: `${category}-${curr}`,
-            children: [],
-          });
-        }
-        return accum;
-      }, []);
-    return { title: capitalize(category), level: 2, slug: category, children: subcategories };
-  });
+  if (filteredTokens.value === null) return [];
+  filteredHeaders.value = updateHeadersRecursively(filteredTokens.value, null);
+};
+
+const updateHeadersRecursively = (node, category) => {
+  return Object.keys(node)
+    .filter(subNodeKey => subNodeKey !== '_children')
+    .map(subNodeKey => {
+      return {
+        title: capitalize(subNodeKey),
+        level: 2,
+        slug: category === null ? subNodeKey : `${category}-${subNodeKey}`,
+        children: updateHeadersRecursively(node[subNodeKey], category === null ? subNodeKey : category),
+      };
+    });
 };
 
 const setFormat = (newFormat) => {
@@ -145,10 +144,10 @@ const filterTokens = () => {
   const regexArray = searchRegexArray.map(searchRegex => new RegExp(searchRegex, 'i'));
 
   filteredTokens.value = filterTokenNode(processedTokens[format.value][theme.value], null, regexArray);
+  console.log(filteredTokens.value);
   updateHeaders();
 };
 
-// eslint-disable-next-line complexity
 const filterTokenNode = (node, name, regexArray) => {
   const subNodesRet = {};
   let childrenRet = [];
@@ -158,14 +157,17 @@ const filterTokenNode = (node, name, regexArray) => {
     if (regex.test(name)) matchingRegex.push(regex);
     else nonMatchingRegex.push(regex);
   });
+  // If the name matches all the terms, return the entire node
   if (nonMatchingRegex.length === 0) {
     return node;
+    // else return the children that match
   } else if (node._children) {
     childrenRet = node._children.filter(token => !token.hidden).filter(token => {
       const { name: tokenName, tokenValue, keywords } = token;
       return nonMatchingRegex.every(regex => regex.test(tokenName) || regex.test(tokenValue) || regex.test(keywords));
     });
   }
+  // look for other categories in the same level and filter recursively
   Object.keys(node).forEach(key => {
     if (key !== '_children') {
       const subNode = filterTokenNode(node[key], key, nonMatchingRegex);
@@ -174,7 +176,7 @@ const filterTokenNode = (node, name, regexArray) => {
   });
   if (Object.keys(subNodesRet).length > 0 || childrenRet.length > 0) {
     return {
-      ...(name !== null ? { _children: childrenRet } : null),
+      _children: childrenRet,
       ...subNodesRet,
     };
   }
@@ -190,11 +192,9 @@ const resetSearch = () => {
   setSearchCriteria();
 };
 
-const filteredTokensKeys = computed(() => Object.keys(filteredTokens.value));
-
 const hasSearchTerm = computed(() => searchInput.value && searchInput.value.trim().length > 0);
 
-const noSearchResults = computed(() => filteredTokensKeys.value.length === 0);
+const noSearchResults = computed(() => filteredTokens.value === null);
 
 /*
 Before mount process the file tokensJson and fill processedTokens with the data we want to show.
@@ -221,7 +221,7 @@ exampleName for the tokens in every format, theme and category.
           surface: {...},
           ...
         typography: {...},
-        shadow: [...],
+        shadow: {...},
         size: {...},
         space: {...},
         component: {...}
@@ -258,142 +258,8 @@ const addTokens = () => {
   });
 };
 
-// eslint-disable-next-line complexity
-const addTokenToStructure = (token, format, structure) => {
-  const [key, value] = token;
-  if (!value[FORMAT_MAP[format]] || !value[FORMAT_MAP.CSS]) return;
-
-  const { name, value: tokenValue, description, keywords } = value[FORMAT_MAP[format]];
-  const { value: exampleValue, name: exampleName } = value[FORMAT_MAP.CSS];
-  const displayToken = { exampleValue, exampleName, name, tokenValue, description, keywords };
-
-  if (isBaseToken(key)) return;
-
-  const splitKeys = key.split('/');
-
-  // COLOR
-  if (key.startsWith('color') && SUBCATEGORY_MAP.color.includes(splitKeys[1])) {
-    addTokenToSubcategory(displayToken, 'color', SUBCATEGORY_MAP.color.find(sub => sub === splitKeys[1]), structure);
-    return;
-  }
-
-  if (key.startsWith('color/brand') || key.startsWith('color/gradient')) {
-    structure.color.brand._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('opacity')) {
-    structure.color.opacity._children.push(displayToken);
-    return;
-  }
-
-  if ((CATEGORY_MAP.component.includes(splitKeys[0]) && (splitKeys[1] === 'color')) ||
-    (CATEGORY_MAP.component.includes(splitKeys[1]) && splitKeys[0] === 'theme')) {
-    structure.color.components._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('color')) {
-    structure.color.base._children.push(displayToken);
-  }
-
-  // TYPOGRAPHY
-  if (key.startsWith('typography')) {
-    structure.typography['font style']._children.push({ ...displayToken, hidden: true });
-    return;
-  }
-
-  if (key.startsWith('font') && SUBCATEGORY_MAP.font.includes(splitKeys[1])) {
-    const displaySubcategory = `font ${SUBCATEGORY_MAP.font.find(sub => sub === splitKeys[1])}`;
-    addTokenToSubcategory(displayToken, 'typography', displaySubcategory, structure);
-    return;
-  }
-
-  if (key.startsWith('font/text-case')) {
-    structure.typography.textcase._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('font/lineHeight')) {
-    structure.typography['line height']._children.push(displayToken);
-    return;
-  }
-
-  if (CATEGORY_MAP.component.includes(splitKeys[0]) &&
-    (splitKeys[1] === 'font' || splitKeys[1] === 'lineHeight')) {
-    structure.typography.components._children.push(displayToken);
-    return;
-  }
-
-  // SHADOW
-  if (key.startsWith('shadow')) {
-    structure.shadow._children.push({ ...displayToken, hidden: true });
-    return;
-  }
-
-  // SIZE
-  if (key.startsWith('size') && key.endsWith('negative')) {
-    structure.size.negative._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('size') && key.endsWith('percent')) {
-    structure.size.percentage._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('size') && SUBCATEGORY_MAP.size.includes(splitKeys[1])) {
-    addTokenToSubcategory(displayToken, 'size', SUBCATEGORY_MAP.size.find(sub => sub === splitKeys[1]), structure);
-    return;
-  }
-
-  if (key.startsWith('size/radius')) {
-    structure.size.radius._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('size/border')) {
-    structure.size.border._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('size')) {
-    structure.size._children.push(displayToken);
-    return;
-  }
-
-  if (CATEGORY_MAP.component.includes(splitKeys[0]) && splitKeys[1] === 'size') {
-    if (splitKeys[2] === 'border') {
-      structure.size.components._children.push({ ...displayToken, hidden: true });
-    } else {
-      structure.size.components._children.push(displayToken);
-    }
-    return;
-  }
-
-  // SPACE
-  if (key.startsWith('space') && key.endsWith('negative')) {
-    structure.space.negative._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('space') && key.endsWith('percent')) {
-    structure.space.percentage._children.push(displayToken);
-    return;
-  }
-
-  if (key.startsWith('space')) {
-    structure.space._children.push(displayToken);
-  }
-};
-
-const addTokenToSubcategory = (token, category, subcategory, structure) => {
-  structure[category][subcategory]._children.push(token);
-};
-
 /*
 * Tokens that are a combination of other tokens.
-* Adds property "hidden" to the base tokens, because we don't want to show them on the list.
 * Only apply for the categories typography and shadow, and only in CSS format.
 */
 const addComposedTokens = () => {
@@ -412,8 +278,6 @@ const addComposedTokens = () => {
     ];
   });
 };
-
-const isBaseToken = (name) => name.endsWith('base') || name.endsWith('root');
 </script>
 
 <style scoped>
